consolidate model schemas
