+++ backend/pkg/database/models/chat_message.go
@@ -7,9 +7,9 @@ import (
 	"gorm.io/datatypes"
 )
 
-// ChatMessageModel is the GORM model for chat messages
+// ChatMessage is the GORM model for chat messages
 // Note: This stays close to the database as it's used by core/chat
-type ChatMessageModel struct {
+type ChatMessage struct {
 	ID        uuid.UUID      `json:"id" gorm:"type:uuid;primaryKey;default:uuid_generate_v4()"`
 	ArticleID uuid.UUID      `json:"article_id" gorm:"type:uuid;not null;index"`
 	Role      string         `json:"role" gorm:"not null"`
@@ -18,6 +18,6 @@ type ChatMessageModel struct {
 	CreatedAt time.Time      `json:"created_at" gorm:"autoCreateTime"`
 }
 
-func (ChatMessageModel) TableName() string {
+func (ChatMessage) TableName() string {
 	return "chat_message"
 }
