+++ backend/pkg/database/models/tag.go
@@ -6,19 +6,19 @@ import (
 	"backend/pkg/core/tag"
 )
 
-// TagModel is the GORM model for tags
-type TagModel struct {
+// Tag is the GORM model for tags
+type Tag struct {
 	ID        int       `json:"id" gorm:"primaryKey;autoIncrement"`
 	Name      string    `json:"name" gorm:"uniqueIndex;not null"`
 	CreatedAt time.Time `json:"created_at" gorm:"autoCreateTime"`
 }
 
-func (TagModel) TableName() string {
+func (Tag) TableName() string {
 	return "tag"
 }
 
 // ToCore converts the GORM model to the domain type
-func (m *TagModel) ToCore() *tag.Tag {
+func (m *Tag) ToCore() *tag.Tag {
 	return &tag.Tag{
 		ID:        m.ID,
 		Name:      m.Name,
@@ -26,9 +26,9 @@ func (m *TagModel) ToCore() *tag.Tag {
 	}
 }
 
-// TagModelFromCore creates a GORM model from the domain type
-func TagModelFromCore(t *tag.Tag) *TagModel {
-	return &TagModel{
+// TagFromCore creates a GORM model from the domain type
+func TagFromCore(t *tag.Tag) *Tag {
+	return &Tag{
 		ID:        t.ID,
 		Name:      t.Name,
 		CreatedAt: t.CreatedAt,
